refactor inference engines
